fix(production): 실생산량 정수나눗셈 버그 수정 및 생산/대기 큐 분리
- OrderService.java: actualQty 계산 시 shortage를 (double)로 캐스트하여 정수 나눗셈 방지 (부족분 2개, 수율 0.9 → 실생산량 3이 되던 버그 수정)
- ProductionService.java: getQueue()를 startedAt==null인 대기 항목만 반환하도록 변경하여 현재 생산 중 항목이 대기 큐에 중복 표시되던 문제 수정, 취소용 getAllEntries() 추가
- ProductionController.java: handleCancel()에서 getAllEntries() 사용으로 현재 생산 중 항목도 취소 목록에 포함
- ProductionServiceTest.java: getQueue() 의미 변경에 맞춰 enqueueWaitsWhenBusy, cancelWaitingEntry 테스트 기댓값 수정

+++ src/main/java/org/example/sampleordersystem/service/OrderService.java
@@ -55,7 +55,7 @@ public class OrderService {
         } else {
             order.transitionTo(OrderStatus.PRODUCING);
             orderRepository.save(order);
-            int actualQty = (int) Math.ceil(shortage / (sample.getYield() * 0.9));
+            int actualQty = (int) Math.ceil((double) shortage / (sample.getYield() * 0.9));
             double totalMinutes = sample.getAvgProductionMinutes() * actualQty;
             ProductionEntry entry = new ProductionEntry(
                 order.getOrderId(), sample.getId(), shortage, actualQty, totalMinutes, null
